requirement.txt
Django==1.7.1
Pillow==2.6.1
South==1.0.1
Unidecode==0.04.16
argparse==1.1
cmsplugin-filer==0.10.1
coverage==3.7.1
django-allauth==0.18.0
django-angular==0.7.10
django-annoying==0.8.0
django-appconf==0.6
django-classy-tags==0.5
django-cms==3.0.7
django-compressor==1.4
django-crispy-forms==1.4.0
django-debug-panel==0.8.0
django-debug-toolbar==1.2.2
django-extensions==1.4.8
django-factory==0.11
django-filer==0.9.8
django-migration-fixture==0.1.4
django-model-utils==2.2
django-modeltranslation==0.8
django-mptt==0.6.1
django-phonenumber-field==0.7.2
django-photologue==3.1.1
django-polymorphic==0.6
django-reversion==1.8.0
django-scheduler==0.7.3
django-sekizai==0.7
django-sortedm2m==0.8.1
django-tastypie==0.12.1
django-tastypie-swagger==0.1.2
django-treebeard==3.0
djangocms-admin-style==0.2.2
djangocms-picture==0.1
djangocms-text-ckeditor==2.4.2
easy-thumbnails==2.2
extras==0.0.3
factory-boy==2.5.2
fake-factory==0.5.0
googlemaps==2.0
html5lib==0.999
linecache2==1.0.0
oauthlib==0.7.2
phonenumbers==7.0.4
psycopg2==2.6
pygeocoder==1.2.5
python-dateutil==2.3
python-mimeparse==0.1.4
python-openid==2.2.5
pytz==2014.10
qrcode==5.1
requests==2.5.0
requests-oauthlib==0.4.2
six==1.8.0
sqlparse==0.1.14
testtools==1.7.1
traceback2==1.4.0
unittest2==1.0.1
vobject==0.8.1c
wsgiref==0.1.2
